+++ app/component/password/information_sidebar_section.js
@@ -11,6 +11,7 @@
  * @link          https://www.passbolt.com Passbolt(tm)
  * @since         2.0.0
  */
+import MadBus from 'passbolt-mad/control/bus';
 import SecondarySidebarSectionComponent from 'app/component/workspace/secondary_sidebar_section';
 
 import template from 'app/view/template/component/password/information_sidebar_section.stache!';
@@ -39,6 +40,31 @@ var InformationSidebarSectionComponent = SecondarySidebarSectionComponent.extend
 	 */
 	'{resource} updated': function (item) {
 		this.refresh();
+	},
+
+	/**
+	 * The password has been clicked.
+	 * @param {HTMLElement} el The element the event occurred on
+	 * @param {HTMLEvent} ev The event which occurred
+	 */
+	' li.password .secret-copy > a click': function (el, ev) {
+		// Get secret out of Resource object.
+		var secret = this.options.resource.secrets[0].data;
+		// Request decryption. (delegated to plugin).
+		MadBus.trigger('passbolt.secret.decrypt', secret);
+	},
+
+	/**
+	 * The username has been clicked.
+	 * @param {HTMLElement} el The element the event occurred on
+	 * @param {HTMLEvent} ev The event which occurred
+	 */
+	' li.username .value > a click': function (el, ev) {
+		var username = this.options.resource.username;
+		MadBus.trigger('passbolt.clipboard', {
+			name: 'username',
+			data: username
+		});
 	}
 
 });