+++ app/view/component/password/password_secondary_sidebar.js
@@ -28,17 +28,6 @@ var PasswordSecondarySidebarView = SecondarySidebarView.extend('passbolt.view.co
 	 */
 	'h2 click': function (el, ev) {
 		$(el).next('p').toggle();
-	},
-
-	/**
-	 * Observe when the user clicks on any h2 element, rolldown the following p tag
-	 * @param {HTMLElement} el The element the event occurred on
-	 * @param {HTMLEvent} ev The event which occurred
-	 */
-	'li.password .secret-copy > a click': function (el, ev) {
-		ev.stopPropagation();
-		ev.preventDefault();
-		$(this.element).trigger('password_clicked', [ev]);
 	}
 });